backend/requirements.txt
fastapi
uvicorn
langchain-google-genai
python-dotenv
google-generativeai
langchain
pypdf
openpyxl
pandas
pyinstaller
